+++ ts/examples/chat/src/memory/common.ts
@@ -172,6 +172,35 @@ export function argChunkSize(defaultValue?: number | undefined): ArgDef {
     };
 }
 
+export function recordFromArgs(
+    args: NamedArgs,
+    metadata?: CommandMetadata,
+): Record<string, string> {
+    const record: Record<string, string> = {};
+    const keys = Object.keys(args);
+    for (const key of keys) {
+        const value = args[key];
+        if (typeof value !== "function") {
+            record[key] = value;
+        }
+    }
+    if (metadata !== undefined) {
+        if (metadata.args) {
+            removeKeysFromRecord(record, Object.keys(metadata.args));
+        }
+        if (metadata.options) {
+            removeKeysFromRecord(record, Object.keys(metadata.options));
+        }
+    }
+    return record;
+}
+
+function removeKeysFromRecord(record: Record<string, string>, keys: string[]) {
+    for (const key of keys) {
+        delete record[key];
+    }
+}
+
 export function argToDate(value: string | undefined): Date | undefined {
     return value ? dateTime.stringToDate(value) : undefined;
 }
